de female endings for authorship, etc.
Dear Zotero team,
thank you for your great work! 

There is the need to gernder authorship, etc. in German. It seems odd if there are only male authors, programmers, etc. Therefore it should be changed. 
In the case of "Gast" (guest) I stayed with the original, as it is often used and percieved neutral. Another neutral version can be "geladene Person", which means literally invited person. But "Gästin" could be used as female version of Gast -> "Gast/Gästin".  As a sponsor can be a person but also an organization I did not changed it as there would be a differentiation needed between persons and organizations as sponsors.

Thank you and cheers!